requirements.txt
transformers
datasets
seqeval
accelerate
pandas
numpy
scikit-learn
matplotlib
torch
pytorch-crf
gensim
tqdm
